перевел на MVVM
