+++ app/src/main/java/com/gb/stopwatch/viewmodels/MainActivityViewModel.kt
@@ -1,4 +1,31 @@
 package com.gb.stopwatch.viewmodels
 
-class MainActivityViewModel {
+import com.gb.stopwatch.MainActivityContract
+import com.gb.stopwatch.domain.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+
+class MainActivityViewModel: MainActivityContract.ViewModel()
+{
+    override val timestampProvider = object : TimestampProvider {
+        override fun getMilliseconds(): Long {
+            return System.currentTimeMillis()
+        }
+    }
+
+    override val stopwatchListOrchestrator = StopwatchListOrchestrator(
+        StopwatchStateHolder(
+            StopwatchStateCalculator(
+                timestampProvider,
+                ElapsedTimeCalculator(timestampProvider)
+            ),
+            ElapsedTimeCalculator(timestampProvider),
+            TimestampMillisecondsFormatter()
+        ),
+        CoroutineScope(
+            Dispatchers.Main
+                    + SupervisorJob()
+        )
+    )
 }
